% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{amqp_acks}
\alias{amqp_acks}
\alias{amqp_ack}
\alias{amqp_nack}
\title{Acknowledge or Reject Incoming Messages}
\usage{
amqp_ack(conn, delivery_tag, multiple = FALSE)

amqp_nack(conn, delivery_tag, multiple = FALSE, requeue = FALSE)
}
\arguments{
\item{conn}{An object returned by \code{\link{amqp_connect}}.}

\item{delivery_tag}{The message's numeric identifier.}

\item{multiple}{When \code{TRUE}, (n)ack messages up-to-and-including this
\code{delivery_tag}. By default, we only (n)ack a single message.}

\item{requeue}{When \code{TRUE}, ask the server to requeue the message.
Otherwise, messages are discarded or dead-lettered.}
}
\description{
Notify the server that a message (or a series of messages) have been received
by acknowledging (ack-ing) them. Or, reject incoming messages that you cannot
handle correctly by "nack"-ing them.
}
