% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{amqp_bindings}
\alias{amqp_bindings}
\alias{amqp_bind_queue}
\alias{amqp_unbind_queue}
\title{Bind Queues to Exchanges}
\usage{
amqp_bind_queue(conn, queue, exchange, routing_key = "", ...)

amqp_unbind_queue(conn, queue, exchange, routing_key = "", ...)
}
\arguments{
\item{conn}{An object returned by \code{\link{amqp_connect}}.}

\item{queue}{The name of the queue.}

\item{exchange}{The name of an exchange.}

\item{routing_key}{The routing key to use (if applicable).}

\item{...}{Binding arguments, if any.}
}
\description{
In order for messages to be routed to a queue, the queue must be bound to
the exchange, although the nature of this binding depends on the type of
exchange.
}
