% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{amqp_connect}
\alias{amqp_connect}
\title{Open a Connection to a RabbitMQ Server}
\usage{
amqp_connect(host = "localhost", port = 5672L, vhost = "/",
  username = "guest", password = "guest", timeout = 10L)
}
\arguments{
\item{host}{The server host.}

\item{port}{The server port.}
}
\value{
An external pointer to a connection.
}
\description{
Open a Connection to a RabbitMQ Server
}
