% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{amqp_declare_queue}
\alias{amqp_declare_queue}
\title{Declare a Queue}
\usage{
amqp_declare_queue(conn, queue, quietly = FALSE)
}
\arguments{
\item{conn}{An object returned by \code{\link{amqp_connect}}.}

\item{queue}{The name of a queue.}

\item{quietly}{Suppress queue information echo.}
}
\description{
Declare a Queue
}
