% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{amqp_declare_tmp_queue}
\alias{amqp_declare_tmp_queue}
\title{Declare a Temporary Queue}
\usage{
amqp_declare_tmp_queue(conn, passive = FALSE, exclusive = TRUE)
}
\arguments{
\item{conn}{An object returned by \code{\link{amqp_connect}}.}

\item{exclusive}{}
}
\value{
The name of the temporary queue.
}
\description{
Declare a Temporary Queue
}
