% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{amqp_delete_queue}
\alias{amqp_delete_queue}
\title{Delete a Queue}
\usage{
amqp_delete_queue(conn, queue, if_unused = FALSE, if_empty = FALSE)
}
\arguments{
\item{conn}{An object returned by \code{\link{amqp_connect}}.}

\item{queue}{The name of a queue.}

\item{if_unused}{Delete the queue only if it is unused.}

\item{if_empty}{Delete the queue only if it is empty.}
}
\value{
The number of messages in the queue when it was deleted, invisibly.
}
\description{
Delete a Queue
}
