% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{amqp_get}
\alias{amqp_get}
\title{Get a Message from a Queue}
\usage{
amqp_get(conn, queue, no_ack = FALSE)
}
\arguments{
\item{conn}{An object returned by \code{\link{amqp_connect}}.}

\item{queue}{The name of a queue.}

\item{no_ack}{When \code{TRUE}, tell the server not to expect that messages
will be acknowledged.}
}
\value{
A string containing the message, or a zero-length character vector if
  there is no message in the queue. Messages may have additional properties
  (such as the content type) attached to them as attributes.
}
\description{
Get a message from a given queue.
}
\seealso{
\code{\link{amqp_consume}} for handling messages with a callback.
}
