% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{amqp_publish}
\alias{amqp_publish}
\title{Publish a Message to a Queue}
\usage{
amqp_publish(conn, queue, msg, content_type = "text/plain")
}
\arguments{
\item{conn}{An object returned by \code{\link{amqp_connect}}.}

\item{queue}{The name of a queue.}

\item{msg}{The message to send.}

\item{content_type}{The message content type.}
}
\description{
Publish a Message to a Queue
}
