% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{amqp_reconnect}
\alias{amqp_reconnect}
\title{Restablish a Connection to a RabbitMQ Server}
\usage{
amqp_reconnect(conn)
}
\arguments{
\item{conn}{An object returned by \code{\link{amqp_connect}}.}
}
\description{
Restablish a Connection to a RabbitMQ Server
}
