% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{amqp_disconnect}
\alias{amqp_disconnect}
\title{Close a Connection to a RabbitMQ Server}
\usage{
amqp_disconnect(conn)
}
\arguments{
\item{conn}{An external pointer to a connection.}
}
\description{
Close a Connection to a RabbitMQ Server
}
