% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{amqp_get}
\alias{amqp_get}
\title{Get a Message from a Queue}
\usage{
amqp_get(conn, queue)
}
\arguments{
\item{conn}{An object returned by \code{\link{amqp_connect}}.}

\item{queue}{The name of a queue.}
}
\value{
A string containing the message, or a zero-length character vector
  if there is no message in the queue.
}
\description{
Get a Message from a Queue
}
