% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{amqp_publish}
\alias{amqp_publish}
\title{Publish a Message to an Exchange}
\usage{
amqp_publish(conn, routing_key, body, exchange = "",
  content_type = "text/plain", mandatory = FALSE, immediate = FALSE)
}
\arguments{
\item{conn}{An object returned by \code{\link{amqp_connect}}.}

\item{routing_key}{The routing key for the message. For the default
exchange, this is the name of a queue.}

\item{body}{The message to send.}

\item{exchange}{The exchange to route the message through.}

\item{content_type}{The message content type.}

\item{mandatory}{When \code{TRUE}, demand that the message is placed in a
queue.}

\item{immediate}{When \code{TRUE}, demand that the message is delivered
immediately.}
}
\description{
Publishes a message to an exchange with a given routing key.
}
